feat: update VeinMiner to latest version
VeinMiner has own maven repository, but it is not secure (http), so I built VeinMiner locally and added files to the `libs` directory.
Additionally, VeinMiner has many breaking changes between v1 and v2, so I have to update VeinMinerBridgeImpl. I have tried to maintain as much functionality as possible, but I am not sure if it actually works.